backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
rasterio==1.4.3
geopandas==1.0.1
numpy==2.2.1
Pillow==11.1.0
python-dotenv==1.0.1
pyproj==3.7.0
shapely==2.0.6
